Add comments and added validation so only staff members can perform some of the views that were forgotten

+++ booking/views.py
@@ -29,6 +29,12 @@ def error_404(request, exception):
 
 @login_required
 def place_booking(request):
+    """
+    Views the booking form and checks that the
+    input is valid before submitting. If it is
+    not, a message will show and redirect back
+    to the form.
+    """
     if request.method == 'POST':
         booking = PlaceBooking(user=request.user)
         form = BookingForm(request.POST, instance=booking)
@@ -37,6 +43,11 @@ def place_booking(request):
             messages.success(request, 'Booking placed successfully.\
                  Please allow up to 24 hours for callback.')
             return redirect('place_booking')
+        else:
+            messages.error(
+                request, 'Review must contain valid text. Try again.'
+                )
+            return redirect('place_booking')
     else:
         form = BookingForm()
     return render(request, 'booking/place_booking.html', {'form': form})
@@ -44,6 +55,10 @@ def place_booking(request):
 
 @login_required
 def view_booking(request):
+    """
+    Lets staff member view all bookings, and
+    regular user only his own bookings.
+    """
     if request.user.is_staff:
         bookings = PlaceBooking.objects.all()
     else:
@@ -56,26 +71,38 @@ def view_booking(request):
 
 @login_required
 def delete_booking(request, booking_id):
+    """
+    Deletes booking and sends an email to the user who placed it.
+    John at tutor support helped med figure out this one.
+    """
     email_to = None
-    try:
-        booking = get_object_or_404(PlaceBooking, id=booking_id)
-        email_to = booking.email
-        subject = 'Your booking'
-        message = f'Hi {booking.first_name}, your booking on\
-             {booking.date_for_visit} has been cancelled.'
-        email_from = 'bestlasbooking@gmail.com'
-        recipient_list = [email_to, ]
-        send_mail(subject, message, email_from, recipient_list)
-        booking.delete()
-        messages.success(request, 'Booking deleted successfully.')
-        return redirect('my_account')
-    except Http404 as err:
-        messages.error(request, 'Oops, booking not found.')
-        return redirect('my_account')
+    if request.user.is_staff:
+        try:
+            booking = get_object_or_404(PlaceBooking, id=booking_id)
+            email_to = booking.email
+            subject = 'Your booking'
+            message = f'Hi {booking.first_name}, your booking on\
+                {booking.date_for_visit} has been cancelled.'
+            email_from = 'bestlasbooking@gmail.com'
+            recipient_list = [email_to, ]
+            send_mail(subject, message, email_from, recipient_list)
+            booking.delete()
+            messages.success(request, 'Booking deleted successfully.')
+            return redirect('my_account')
+        except Http404 as err:
+            messages.error(request, 'Oops, booking not found.')
+            return redirect('my_account')
+    else:
+        return redirect('home')
 
 
 @login_required
 def edit_booking(request, booking_id):
+    """
+    Presents the booking form with the recent informtion added.
+    When update is done the user gets an email that an update
+    has been made.
+    """
     email_to = None
     try:
         booking = get_object_or_404(PlaceBooking, id=booking_id)
@@ -107,24 +134,37 @@ def edit_booking(request, booking_id):
 
 @login_required
 def approve_booking(request, booking_id):
-    booking = get_object_or_404(PlaceBooking, id=booking_id)
-    booking.approved = not booking.approved
-    booking.save()
-    email_to = booking.email
-    subject = 'Your booking'
-    message = f'Hi {booking.first_name}, your booking on\
-        {booking.date_for_visit} has been updated.\
-        Log in to see details.\
-        Log in: https://locksmith-booking.herokuapp.com/'
-    email_from = 'bestlasbooking@gmail.com'
-    recipient_list = [email_to, ]
-    send_mail(subject, message, email_from, recipient_list)
-    messages.success(request, 'Booking Updated successfully!')
-    return redirect('my_account')
+    """
+    Let staff members approve a booking
+    or withdraw an approved booking.
+    User gets an email that a update has been made.
+    """
+    if request.user.is_staff:
+        booking = get_object_or_404(PlaceBooking, id=booking_id)
+        booking.approved = not booking.approved
+        booking.save()
+        email_to = booking.email
+        subject = 'Your booking'
+        message = f'Hi {booking.first_name}, your booking on\
+            {booking.date_for_visit} has been updated.\
+            Log in to see details.\
+            Log in: https://locksmith-booking.herokuapp.com/'
+        email_from = 'bestlasbooking@gmail.com'
+        recipient_list = [email_to, ]
+        send_mail(subject, message, email_from, recipient_list)
+        messages.success(request, 'Booking Updated successfully!')
+        return redirect('my_account')
+    else:
+        return redirect('home')
 
 
 @login_required
 def place_review(request):
+    """
+    Presents the review form if user is logged in.
+    Checks that it is valid and throws a message
+    if it is not.
+    """
     if request.method == 'POST':
         review = Review(user=request.user)
         form = ReviewForm(request.POST, instance=review)
@@ -150,6 +190,9 @@ def place_review(request):
 
 @login_required
 def view_review(request):
+    """
+    Let staff members view all reviews.
+    """
     if request.user.is_staff:
         reviews = Review.objects.all()
         context = {
@@ -162,6 +205,10 @@ def view_review(request):
 
 @login_required
 def delete_review(request, review_id):
+    """
+    Let staff members delete a review.
+    If it does not exist, a message is shown.
+    """
     if request.user.is_staff:
         try:
             review = get_object_or_404(Review, id=review_id)
@@ -177,6 +224,9 @@ def delete_review(request, review_id):
 
 @login_required
 def view_users(request):
+    """
+    Let staff memebers view all the users.
+    """
     if request.user.is_staff:
         users = User.objects.all()
         context = {
@@ -189,6 +239,11 @@ def view_users(request):
 
 @login_required
 def delete_user(request, user_id):
+    """
+    Let staff member delete a user.
+    If the user does not exist, a message
+    will show.
+    """
     if request.user.is_staff:
         try:
             user = get_object_or_404(User, id=user_id)
@@ -199,27 +254,42 @@ def delete_user(request, user_id):
             messages.error(request, 'Oops, user not found.')
             return redirect('users')
     else:
+        messages.error(request, 'You do not have permission to do this.')
         return redirect('home')
 
 
 @login_required
 def toggle_staff(request, user_id):
-    user = get_object_or_404(User, id=user_id)
-    user.is_staff = not user.is_staff
-    user.save()
-    email_to = user.email
-    subject = 'Your status has changed'
-    message = f'Hi {user.username}, your authorization has changed.\
-        Log in: https://locksmith-booking.herokuapp.com/'
-    email_from = 'bestlasbooking@gmail.com'
-    recipient_list = [email_to, ]
-    send_mail(subject, message, email_from, recipient_list)
-    messages.success(request, 'User Updated successfully!')
-    return redirect('users')
+    """
+    Make it possible to give user staff status,
+    or remove it. User gets an email that a
+    change has been made.
+    """
+    if request.user.is_staff:
+        user = get_object_or_404(User, id=user_id)
+        user.is_staff = not user.is_staff
+        user.save()
+        email_to = user.email
+        subject = 'Your status has changed'
+        message = f'Hi {user.username}, your authorization has changed.\
+            Log in: https://locksmith-booking.herokuapp.com/'
+        email_from = 'bestlasbooking@gmail.com'
+        recipient_list = [email_to, ]
+        send_mail(subject, message, email_from, recipient_list)
+        messages.success(request, 'User Updated successfully!')
+        return redirect('users')
+    else:
+        messages.error(request, 'You do not have permission for this.')
+        return redirect('home')
 
 
 @login_required
 def approve_review(request, review_id):
+    """
+    Make it possible to approve a review so it will
+    show on the home page. Or remove a review from
+    the home page.
+    """
     review = get_object_or_404(Review, id=review_id)
     review.approved = not review.approved
     review.save()
